move role/scope fixtures to conftest
so they can be more easily reused
