+++ jupyterhub/tests/test_roles.py
@@ -13,10 +13,8 @@ from traitlets.config import Config
 from .. import orm
 from .. import roles
 from ..scopes import get_scopes_for
-from ..utils import maybe_future
 from ..utils import utcnow
 from .mocking import MockHub
-from .test_scopes import create_temp_role
 from .utils import add_user
 from .utils import api_request
